v3.6.0-rc.2: split index.ts (3665 lines) + add tool-manifest.ts
v3.6.0 sprint Phase 2 of 4. Pure refactor: same CLI surface, same
registered tools, 712 tests pass. Closes the second monolith file
flagged by the audit cycle (after tools.ts in rc.1).

Structure:
- src/index.ts          (84 lines)   slim entry: VERSION + CLI guard
                                     + re-exports for v3.5.x surface
- src/cli.ts            (702 lines)  main() + commander program (all
                                     12 subcommands)
- src/server.ts         (877 lines)  MCP server construction: ServeOpts,
                                     ServerDeps, prepareServerDeps,
                                     buildMcpServer, startServer,
                                     formatReadyBanner, buildEmbedText,
                                     sync{EmbedDb,FtsIndex,Pdf*}
- src/tool-registry.ts  (1300 lines) registerReadTools/WriteTools/
                                     FtsTools/Resources/ChunkResource +
                                     utility helpers (textResult,
                                     encodeNotePath, decodeNotePath,
                                     parsePositiveInt,
                                     parseQuantizationMode, embedDbPath)
- src/prompts.ts        (790 lines)  registerPrompts: 19 MCP prompts
- src/tool-manifest.ts  (318 lines)  NEW machine-readable registry of
                                     all 44 tools (kind + gating +
                                     summary). Single source of truth.

Tool-manifest breakdown (matches count math invariant):
  33 read (always-on) + 1 fts (--persistent-index +
  --diagnostic-search-tools) + 3 diagnostic
  (--diagnostic-search-tools) + 7 write (--enable-write) = 44 total.

tests/docs-consistency.test.ts pivoted to read TOOL_MANIFEST directly
instead of regex-parsing source code. Removed an entire class of
test brittleness — refactor-resistant by construction.

Module dependency graph is cycle-safe:
- index.ts → cli.ts (main value)
- cli.ts → index.ts (VERSION literal, evaluated at module-init, safe)
- cli.ts → server.ts → tool-registry.ts (linear)
- tool-registry.ts → server.ts (type-only for ServerDeps, erased at
                                 compile time)

712 tests pass · branches 75.29% · lint clean · tsc strict clean ·
smoke pass · version-consistency green at 3.6.0-rc.2 (5 surfaces) ·
changelog-coverage gate OK.

npm dist-tag: rc (NOT latest). Users on latest stay on v3.5.14.
Install: npm i @oomkapwn/enquire-mcp@rc.

Next: v3.6.0-rc.3 — Full TSDoc on 44 tools + 19 prompts + 20 src/
modules.
